results pagination

+++ Restaurants.Application/Dishes/Commands/DeleteDishesForRestaurant/DeleteDishesForRestaurantCommandHandler.cs
@@ -3,6 +3,7 @@ using Microsoft.Extensions.Logging;
 using Restaurants.Application.Dishes.Queries.GetDishesForRestaurant;
 using Restaurants.Domain.Entities;
 using Restaurants.Domain.Exceptions;
+using Restaurants.Domain.Interfaces;
 using Restaurants.Domain.Repositories;
 using Restaurants.Infrastructure.Repositories;
 
@@ -10,7 +11,8 @@ namespace Restaurants.Application.Dishes.Commands.DeleteDishesForRestaurant;
 
 public class DeleteDishesForRestaurantCommandHandler(ILogger<GetDishesForRestaurantQueryHandler> logger,
 IRestaurantsRepository restaurantsRepository,
-IDishesRepository dishesRepository) 
+IDishesRepository dishesRepository,
+IRestaurantAuthorizationService restaurantAuthorizationService) 
     : IRequestHandler<DeleteDishesForRestaurantCommand>
 {
     public async Task Handle(DeleteDishesForRestaurantCommand request, CancellationToken cancellationToken)
@@ -19,6 +21,9 @@ IDishesRepository dishesRepository)
 
         var restaurant = await restaurantsRepository.GetByIdAsync(request.RestaurantId) 
             ?? throw new NotFoundException(nameof(Restaurant), request.RestaurantId.ToString());
+
+        if (!restaurantAuthorizationService.Authorize(restaurant, Domain.Constants.ResourceOperation.Delete))
+            throw new ForbidException();
 
         await dishesRepository.Delete(restaurant.Dishes);
     }
